+++ src/main/java/com/mitc/intellij/openapi/utils/specs/Format.java
@@ -1,0 +1,5 @@
+package com.mitc.intellij.openapi.utils.specs;
+
+public enum Format {
+    JSON, YML
+}