+++ src/main/java/com/mitc/intellij/openapi/utils/specs/Spec.java
@@ -1,0 +1,5 @@
+package com.mitc.intellij.openapi.utils.specs;
+
+public enum Spec {
+    OAS2, OAS3, UNDEFINED
+}